fix(analytics): fixed build step that broke analytics
