+++ .gitignore
@@ -45,3 +45,4 @@ amplify/backend/amplify-meta.json
 amplify/backend/awscloudformation
 aws-exports.js
 awsconfiguration.json
+credentials.js
