+++ package.json
@@ -80,8 +80,9 @@
     "clean": "yarn delete-cache -- node_modules && yarn",
     "contribute": "scripts/contribute/cli.js",
     "validate-content": "markdownlint content/**/* && scripts/validations.js",
-    "start": "yarn delete-cache && gatsby develop",
-    "build": "gatsby build --prefix-paths",
+    "prep": "yarn delete-cache && scripts/gather-credentials.js",
+    "start": "yarn prep && gatsby develop",
+    "build": "yarn prep && gatsby build --prefix-paths",
     "serve": "gatsby serve",
     "cy:dev": "cypress open --config baseUrl=http://localhost:8000",
     "cy:ci": "cypress run --config baseUrl=http://localhost:9000",
