+++ src/utilities/track.js
@@ -1,21 +1,19 @@
 import Amplify, {Analytics} from 'aws-amplify';
 import {curry} from 'ramda';
-
-const awsmobile =
-  process.env.NODE_ENV === 'production' ? require('~/aws-exports') : null;
+import credentials from '../credentials';
 
 // wiped on page refresh / non-pwa redirects
 let configured = false;
 
 const track = curry((name, attributes) => {
   if (!configured) {
-    awsmobile && Amplify.configure(awsmobile);
+    credentials && Amplify.configure(credentials);
     configured = true;
-    console.log('configured analytics');
+    // console.log('configured analytics');
   }
 
-  awsmobile && Analytics.record({name, attributes});
-  console.log('triggered analytics');
+  credentials && Analytics.record({name, attributes});
+  // console.log('triggered analytics');
 });
 
 export const internalPageView = props =>
